refactor notifications to a new module
